% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericStat.R
\name{dataNumericStats}
\alias{dataNumericStats}
\title{Title}
\usage{
dataNumericStats(df)
}
\arguments{
\item{df}{A data set}
}
\value{
Result of the function summary stats
}
\description{
Title
}
\examples{
dataNumericStats(iris)
}
