% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataOutlaiers.R
\name{dataOutlaiers}
\alias{dataOutlaiers}
\title{Title}
\usage{
dataOutlaiers(df)
}
\arguments{
\item{df}{data set}
}
\value{
Result of the function outlier data points for all columns
}
\description{
Title
}
\examples{
dataOutlaiers(iris)


}
