% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelation.R
\name{getCorrelation}
\alias{getCorrelation}
\title{Title}
\usage{
getCorrelation(df)
}
\arguments{
\item{df}{data set}
}
\value{
Gives Correlation graph of entire data
}
\description{
Title
}
\examples{
getCorrelation(iris)

}
