% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistogram.R
\name{getHistogram}
\alias{getHistogram}
\title{Title}
\usage{
getHistogram(df)
}
\arguments{
\item{df}{data set}
}
\value{
histogram plot for all numerical columns in data set
}
\description{
Title
}
\examples{
getHistogram(iris)


}
