% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{graph.Univariate}
\alias{graph.Univariate}
\title{Generate Graphs for Univariate Analysis.}
\usage{
graph.Univariate(dataf, variable = 1:ncol(dataf), generate.png = FALSE,
generated.file.location = paste(getwd(), '/Images/Univariate'))
}
\arguments{
\item{dataf}{a (non-empty) DataFrame.}

\item{variable}{an optional(non-empty) numeric vector of column numbers.}

\item{generate.png}{an optional (non-empty) boolean value to specify whether
the visualization has to be exported as png or not.}

\item{generated.file.location}{an optional (non-empty) string which specifies
where the exported images needs to be stored.}
}
\value{
NA
}
\description{
This function lets the user perform univariate analysis of every column present in
the dataframe by default. User can specify the particular rows he wants to create
visualizations for by assigning some value to the 'variable' argument. The user can also choose
to export these visualizations in png format to the directory specified by the user.
}
\examples{
## To generate visualization of every column.
graph.Univariate(mtcars)

## To generate visualization of specefic columns.
graph.Univariate(mtcars, variable = c(1,2,3))

## To generate visualization and store it as image.
graph.Univariate(mtcars, variable = c(1,2,3), generate.png = TRUE,
generated.file.location = paste(getwd(), '/Images/Univariate'))

}
