% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingData.R
\name{missingDataNumber}
\alias{missingDataNumber}
\title{Title}
\usage{
missingDataNumber(df)
}
\arguments{
\item{df}{data set}
}
\value{
Result of the function missing data points for all columns
}
\description{
Title
}
\examples{
missingDataNumber(iris)
}
