% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{test.all_cat}
\alias{test.all_cat}
\title{Perform every test for Categorical Target.}
\usage{
test.all_cat(mtcars, 'cyl', 2:ncol(mtcars), tTest = FALSE)
}
\arguments{
\item{dataf}{a (non-empty) dataframe.}

\item{y}{a (non-empty) integer/string which points to the column which will be considered
as the target variable.}

\item{x}{a (non-empty) vector of integers/string which points to the columns which should
be used for the test.}

\item{chiSquare}{an optional (non-empty) variable which lets the user decide whether
Chi-Square test should be performed.}

\item{tTest}{an optional (non-empty) variable which lets the user decide whether
T-test should be performed.}
}
\value{
list of dataframes returned from other tests.
}
\description{
This function lets us perform the following tests when the target is categorical:
- Chi-Square Test
- T-Test (Only when number of categories is 2 in target variable)
}
\examples{
## To perform every test.
test.all_cat(mtcars, 'vs', 3:ncol(mtcars))

## To omit a specefic test.
test.all_cat(mtcars, 'cyl', 2:ncol(mtcars), tTest = FALSE)
}
