% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{test.t_test_cat}
\alias{test.t_test_cat}
\title{Student's t-Test. (Category for 2 values vs Numeric)}
\usage{
test.t_test_cat(dataf, y, x, skipMessage = FALSE)
}
\arguments{
\item{dataf}{a (non-empty) dataframe.}

\item{y}{a (non-empty) integer which points to the column which will be considered
as the target variable.}

\item{x}{a (non-empty) vector of integers which points to the columns which should
be used for the test.}

\item{skipMessage}{an optional (non-empty) boolean which lets the user choose whether
he should get the warning messages.}
}
\value{
dataframe containing the result of the tests.
}
\description{
This function performs t test between the categorical target (upto 2 categories)
and user defined columns of the same dataset.
    Note: Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
}
\examples{
test.t_test_cat(mtcars, 2, 3:ncol(mtcars))
}
