% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{transform.augment_impute}
\alias{transform.augment_impute}
\title{Augments the process of imputation by providing relevant visual cues to the user.}
\usage{
kc_complete <- transform.augment_impute(kc_complete)
}
\arguments{
\item{dataf}{a (non-empty) Dataframe.}

\item{variables}{a list of (non-empty) column names with missing values. all columns
are considered by default.}

\item{missingThreshold}{a (non-empty) proportion of missing values to consider
for imputation lying between 0 to 1.}
}
\value{
dataframe with imputed values.
}
\description{
This function provides the user relevant information about each variable defined
by the user iteratively and lets the user decide the kind of imputation they would
like to make on it. In case the variable is numeric, summary with boxplot and histogram
are displayed whereas in case of catogorical variables, table with bar graph and
pie chart are displayed. User gets the option to impute using 'mean', 'mode' or
'median'.
}
\examples{
# To impute in all the columns with less than 30\% missing values.
transform.augment_impute(kc_complete)

# To impute in specefic columns with less than 10\% missing values.
transform.augment_impute(mtcars, c('mpg', 'cyl'), 0.1)
}
