% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{transform.Group_and_factor}
\alias{transform.Group_and_factor}
\title{Groups columns together according to type.}
\usage{
transform.Group_and_factor(mtcars, convertToFactorFlag = FALSE, conversionThreshold = 5)
}
\arguments{
\item{dataf}{a (non-empty) DataFrame.}

\item{convertToFactorFlag}{an optional (non-empty) boolean value which allows the user
to convert the type of column to factor.}

\item{conversionThreshold}{an optional (non-empty) integer which acts as a threshold
while converting type of column to factor.}
}
\value{
transformed Dataframe
}
\description{
This function goes to through the whole dataframe and groups columns with same
data types together. User has the option to convert the columns into factors
according to the number of unique values present in the column.
}
\examples{
## To transform the dataframe to group columns according to data type.
transform.Group_and_factor(mtcars)

## To convert columns into factors when unique values are less than a specified number.
transform.Group_and_factor(mtcars, convertToFactorFlag = TRUE, conversionThreshold = 3)
}
