% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{transform.impute_option}
\alias{transform.impute_option}
\title{Imputes mean, median or mode in a single column}
\usage{
mtcars <- transform.impute_option(mtcars, 'origin', 'mode')
}
\arguments{
\item{dataf}{a (non-empty) Dataframe}

\item{variable.name}{a (non-empty) column name with missing values}

\item{option}{the method to be used. options - 'mean', 'median' and 'mode'}
}
\value{
dataframe with imputed values
}
\description{
This function imputes the missing data on the column selected by the user,
using the statistic selected by the user
}
\examples{
# To get dataframe with mean imputed in 'wt' column.
transform.impute_option(mtcars, 'wt', 'mean')
}
